+++ server/RdtClient.Data/Enums/TorrentStatus.cs
@@ -6,6 +6,7 @@
         WaitingForFileSelection = 1,
         Downloading = 2,
         Finished = 3,
+        Uploading = 4,
 
         Error = 99
     }