+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.9.5] - 2021-10-28
+### Changed
+- Fixed issues with the simple downloader.
+- Fixed issue with retrying downloads.
+
+### Added
+- Added torrent uploading status from Real Debrid.
+- Restored removing of torrents when deleted from Real Debrid.
+
 ## [1.9.4] - 2021-10-28
 ### Changed
 - Fixed issues retrying torrents after multiple failed downloads in large torrents.